Update README.md to reflect project goals
Updated the README.md so that the goals and general purpose
of the project is outlined and people who visit and read the
document can know and understand what the project is for.

This is also to practice the some good conventions for commit
messages and building the foundational habits for good practices
as a software developer.